// src/app/dashboard/discover/types.ts
export interface Master {
  id: string;
  full_name: string | null;
  avatar_url: string | null;
  specialties: string | null;
  city: string | null;
  country: string | null;
  state: string | null;
  state_code: string | null;
  latitude: number | null;
  longitude: number | null;
  bio: string | null;
}